% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidcid.R
\name{decodeCACTVS}
\alias{decodeCACTVS}
\title{Convert a Base64 encoded Pubchem 881-bit fingerprint to a \code{fingerprint} object}
\usage{
decodeCACTVS(cactvs)
}
\arguments{
\item{cactvs}{A character string containing the Base64 encoded fingerprint}
}
\value{
A \code{fingerprint} object
}
\description{
Pubchem computes 881-bit structural keys using the CACTVS toolkit, which are made
available as Base64 encoded strings. This method converts the Pubchem string to
a \code{fingerprint} object, which can be manipulated using the \code{fingerprint}
package.
}
\examples{
decodeCACTVS('AAADcYBAAAAGAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAGAIAAAAAAAOAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==')

}

